--- src/WritingPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import EntranceContainer from './EntranceContainer';
-
-export default function EntrancePage() {
-
-  return (
-    <EntranceContainer />
-  );
-}
