+++ src/page/EntrancePage.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import EntranceContainer from './EntranceContainer';
+import EntranceContainer from '../container/EntranceContainer';
 
 export default function EntrancePage() {
