+++ src/page/WritePage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import WriteContainer from '../container/WriteContainer';
+
+export default function WritePage() {
+
+  return (
+    <WriteContainer pageIndex={1} />
+  );
+}
