+++ src/presentational/EntranceWritePostcard.jsx
